fix: ensure properties are correctly resolved
Prior to this change only the maven properties were actually resolved
leaving the rest not. This would cause for example `$package$` to end up
being generated as `$package`. Now we ensure everything that we can is
resolved by mimicking the way it's done in the sbt plugin.

NOTE: I still plan on improving this a bit by allowing the user to pass
in values to validate/generate which can be used to generate with
specific values.

Closes #11